app/requirements.txt
mysql-connector-python==8.4.0
Flask==3.0.2